JBR-339 refactor money

+++ src/main/java/com/jbr/middletier/money/reporting/EmailGenerator.java
@@ -49,7 +49,7 @@ public class EmailGenerator {
                                 String username,
                                 String host,
                                 String password,
-                                int weeks) throws EmailGenerationException {
+                                long weeks) throws EmailGenerationException {
         try {
             List<TransactionDTO> emailTransactions = new ArrayList<>();
